+++ util/src/offsetmeasurement.h
@@ -9,7 +9,7 @@ class OffsetMeasurement
 {
 public:
     OffsetMeasurement(int index, int64_t starttime_ns, int64_t endtime_ns,
-                      int samples_sent, int samples, size_t used, int64_t avg_ns, bool valid);
+                      size_t samples_sent, size_t samples, size_t used, int64_t avg_ns, bool valid);
 
     void setOffset_ns(int64_t offset_ns);
     std::string toString() const;
@@ -19,9 +19,9 @@ public:
     int m_index;
     int64_t m_starttime_ns;
     int64_t m_endtime_ns;
-    int m_sentSamples;
-    int m_collectedSamples;
-    int m_usedSamples;
+    size_t m_sentSamples;
+    size_t m_collectedSamples;
+    size_t m_usedSamples;
     int64_t m_offset_ns;
     bool m_valid = true;
     int64_t m_clientOffset_ns = 0;
